move positionValueCalculator
